+++ src/main/java/ru/vladikshk/myRedis/service/WatchdogService.java
@@ -1,7 +1,5 @@
 package ru.vladikshk.myRedis.service;
 
-import java.util.Queue;
-
 public interface WatchdogService {
-    void addForWatch(String key, int millisToExpire);
+    void addForWatch(String key, long millisToExpire);
 }
